save progress

+++ src/main/java/ch/fhnw/therewrite/controller/AnnotationController.java
@@ -1,5 +1,6 @@
 package ch.fhnw.therewrite.controller;
 
+import ch.fhnw.therewrite.AccessHelper;
 import ch.fhnw.therewrite.data.Annotation;
 import ch.fhnw.therewrite.data.Document;
 import ch.fhnw.therewrite.repository.AnnotationRepository;
@@ -8,6 +9,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.parameters.P;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -26,9 +31,12 @@ public class AnnotationController {
         this.annotationRepository = annotationRepository;
         this.documentRepository = documentRepository;
     }
-
+    @PreAuthorize("hasRole('ROLE_USER') or hasRole('ROLE_GUEST')")
     @GetMapping("/all/{documentId}")
-    public ResponseEntity<List<Annotation>> getAnnotationsByDocumentId(@PathVariable String documentId) {
+    public ResponseEntity<List<Annotation>> getAnnotationsByDocumentId(@PathVariable String documentId, @AuthenticationPrincipal UserDetails currentUser) {
+        if(currentUser == null || AccessHelper.verifyUserRights(currentUser.getUsername(), documentId, documentRepository)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+        }
         UUID dId = UUID.fromString(documentId);
         Optional<Document> d = documentRepository.findById(dId);
         if(d.isPresent()) {
@@ -37,16 +45,28 @@ public class AnnotationController {
         return ResponseEntity.status(HttpStatus.EXPECTATION_FAILED).body(null);
     }
 
+    @PreAuthorize("hasRole('ROLE_USER') or hasRole('ROLE_GUEST')")
     @PostMapping("")
-    public Annotation saveAnnotation(@RequestBody Annotation annotation) {
-        return annotationRepository.save(annotation);
+    public Annotation saveAnnotation(@RequestBody Annotation annotation, @AuthenticationPrincipal UserDetails currentUser) {
+        Optional<Annotation> optionalAnno = annotationRepository.findById(annotation.getIdAnnotation());
+        if(optionalAnno.isPresent()) {
+            if (currentUser == null || AccessHelper.verifyUserRights(currentUser.getUsername(), optionalAnno.get().getDocument().getId().toString(), documentRepository)) {
+                return null;
+            }
+            return annotationRepository.save(annotation);
+        }
+        return null;
     }
 
+    @PreAuthorize("hasRole('ROLE_USER') or hasRole('ROLE_GUEST')")
     @Modifying
     @PatchMapping("")
-    public ResponseEntity<Annotation> patchAnnotation(@RequestBody Annotation update) {
+    public ResponseEntity<Annotation> patchAnnotation(@RequestBody Annotation update, @AuthenticationPrincipal UserDetails currentUser) {
         Optional<Annotation> optionalAnno = annotationRepository.findById(update.getIdAnnotation());
         if(optionalAnno.isPresent()) {
+            if(currentUser == null || AccessHelper.verifyUserRights(currentUser.getUsername(), optionalAnno.get().getDocument().getId().toString(), documentRepository)) {
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+            }
             Annotation anno = optionalAnno.get();
             anno.patch(update);
             Annotation resp = annotationRepository.save(anno);
@@ -55,11 +75,15 @@ public class AnnotationController {
         return ResponseEntity.status(HttpStatus.EXPECTATION_FAILED).body(null);
     }
 
+    @PreAuthorize("hasRole('ROLE_USER') or hasRole('ROLE_GUEST')")
     @DeleteMapping("/{annoId}")
-    public ResponseEntity<Annotation> deleteAnnotation(@PathVariable String annoId) {
+    public ResponseEntity<Annotation> deleteAnnotation(@PathVariable String annoId, @AuthenticationPrincipal UserDetails currentUser) {
         UUID aId = UUID.fromString(annoId);
         Optional<Annotation> a = annotationRepository.findById(aId);
         if(a.isPresent()) {
+            if(currentUser == null || AccessHelper.verifyUserRights(currentUser.getUsername(), a.get().getDocument().getId().toString(), documentRepository)) {
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+            }
             Annotation oldAnno = a.get();
             annotationRepository.delete(oldAnno);
             return ResponseEntity.status(HttpStatus.OK).body(oldAnno);
